requirements.txt
# PyYAML
# torch
# torchvision
tqdm
# Pillow
efficientnet_pytorch
pretrainedmodels
# tensorflow
pyyaml
